Move the EditorDocumentHead component out of edit-post
