+++ client/gutenberg/editor/editor-document-head/index.jsx
